test(reassignment): add specs for validate_as_input_for_container
